WIP: Typescriptify
