+++ package.json
@@ -38,7 +38,7 @@
     "style-loader": "^0.22.1",
     "typescript": "^3.0.1",
     "url-loader": "^1.0.1",
-    "webextension-polyfill": "^0.3.0",
+    "webextension-polyfill-ts": "^0.8.7",
     "webpack": "^4.16.5",
     "webpack-cli": "^3.1.0"
   },
