--- src/opelete/browser.js
@@ -1,1 +1,0 @@
-export const browser = browser || chrome;
