+++ src/opelete/ready.ts
@@ -1,4 +1,4 @@
-export default function ready(loaded) {
+export default function ready (loaded: (...args: any[]) => any) {
   if (['interactive', 'complete'].includes(document.readyState)) {
     loaded();
   } else {
