+++ src/popup/components/GeneralSettings.ts
