+++ src/popup/components/OperatorCard.ts
