+++ src/popup/components/OperatorManager.ts
