+++ src/popup/components/__tests__/Checkbox-test.ts
