+++ src/popup/components/__tests__/OperatorCard-test.ts
