+++ src/popup/reducers/storage.ts
